Our paper is working on Face recognition problems.
This paper addresses this challenge by proposing a deep learning framework which is based on MTCNN and FaceNet, which can
recover the canonical view of face images. 
We use the inherent correlation between detection and calibration to improve their performance under the multi-task framework of deep cascading. In particular, we use a three-tiered architecture combined with a well-designed roll neural network algorithm to detect faces and roughly locate key points.
In the FaceNet method, it directly learns the mapping from a face image to a compact Euclidean space, where distance directly corresponds to a measure of facial similarity. Once this space is generated, face recognition, validation and clustering can be easily implemented using the standard FaceNet embedding technique as the feature vector. This approaches dramatically reduce the intra-person variances, while maintaining the interpersonal discriminativeness. Maybe there is something not that
perfect during our experiments, but we are going to summarize our experiments and present some challenges lying ahead in recent face recognition